# backend/api-docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plataforma Scan API Documentation",
    "description": "API documentation for the Plataforma Scan system that provides information for both mobile and web applications of the software.",
    "version": "3.5.0",
    "contact": {
      "name": "Plataforma Scan Support",
      "email": "support@plataformascan.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    },
    {
      "url": "https://api.plataformascan.com",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MANAGER",
              "OPERATOR"
            ]
          },
          "isPending": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Institution": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Polygon": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "coordinates": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "lat": {
                  "type": "number"
                },
                "lng": {
                  "type": "number"
                }
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/admin/operators": {
      "get": {
        "summary": "List all operators",
        "description": "Retrieves a list of all operators",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of operators",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid token"
          },
          "403": {
            "description": "Forbidden - Admin/Manager access required"
          }
        }
      }
    },
...
    "/admin/managers": {
      "get": {
        "summary": "List all managers",
        "description": "Retrieves a list of all managers in the system",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of managers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid token"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        }
      },
      "post": {
        "summary": "Create manager",
        "description": "Creates a new manager account",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password",
                  "verificationCode"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Manager's name"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Manager's email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "Manager's password"
                  },
                  "verificationCode": {
                    "type": "string",
                    "description": "Verification code for manager creation"
                  },
                  "institutionId": {
                    "type": "string",
                    "description": "ID of the institution the manager belongs to (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Manager created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Manager created successfully"
                    },
                    "manager": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or verification code"
          },
          "401": {
            "description": "Unauthorized - Invalid token"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          }
        }
      }
    },
...
    "/institutions": {
      "get": {
        "summary": "List all institutions",
        "description": "Retrieves a list of all institutions",
        "tags": [
          "Institutions"
        ],
        "responses": {
          "200": {
            "description": "List of institutions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Institution"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new institution",
        "description": "Creates a new institution",
        "tags": [
          "Institutions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Institution name"
                  },
                  "address": {
                    "type": "string",
                    "description": "Institution address"
                  },
                  "city": {
                    "type": "string",
                    "description": "Institution city"
                  },
                  "state": {
                    "type": "string",
                    "description": "Institution state"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Institution created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Institution"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized - Invalid token"
          },
          "403": {
            "description": "Forbidden - Admin/Manager access required"
          }
        }
      }
    },
...
    "/polygons": {
      "get": {
        "summary": "List all polygons",
        "description": "Retrieves a list of all polygons",
        "tags": [
          "Polygons"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of polygons",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Polygon"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid token"
          },
          "403": {
            "description": "Forbidden - Admin/Manager access required"
          }
        }
      }
    },
...
    "/google-maps-api-url": {
      "get": {
        "summary": "Get Google Maps API URL",
        "description": "Retrieves the URL for Google Maps API with API key",
        "tags": [
          "Integrations"
        ],
        "responses": {
          "200": {
            "description": "Google Maps API URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "format": "uri"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Google Maps API key not configured"
          }
        }
      }
    },
    "/send-notification": {
      "post": {
        "summary": "Send notification",
        "description": "Sends a push notification to a device",
        "tags": [
          "Notifications"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "playerId",
                  "title",
                  "body"
                ],
                "properties": {
                  "playerId": {
                    "type": "string",
                    "description": "Player ID of the device to send notification to"
                  },
                  "title": {
                    "type": "string",
                    "description": "Notification title"
                  },
                  "body": {
                    "type": "string",
                    "description": "Notification body"
                  },
                  "data": {
                    "type": "object",
                    "description": "Additional data to send with notification"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification sent successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "500": {
            "description": "Failed to send notification"
          }
        }
      }
    },
...
  }
}